requirements.txt
# Function dependencies, for example:
# package>=version
requests
flask
psycopg2-binary